Move undocumented DO methods from easyaccess to do_utils file
